Drop unwired provider options and extract shared resolver trait
