+++ src/Files/File.php
@@ -84,7 +84,7 @@ abstract class File implements HasName, HasProviderOptions
     /**
      * Specify provider-specific options for the file upload. Closures may only capture serializable values.
      *
-     * @param  array<string, mixed>|Closure(Lab|string): array<string, mixed>  $options
+     * @param  array<string, mixed>|Closure(Lab|string): ?array<string, mixed>  $options
      */
     public function withProviderOptions(array|Closure $options): static
     {
