+++ src/PendingResponses/PendingEmbeddingsGeneration.php
@@ -2,7 +2,6 @@
 
 namespace Laravel\Ai\PendingResponses;
 
-use Closure;
 use Illuminate\Contracts\Cache\Repository as CacheRepository;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Traits\Conditionable;
@@ -13,16 +12,16 @@ use Laravel\Ai\Events\ProviderFailedOver;
 use Laravel\Ai\Exceptions\FailoverableException;
 use Laravel\Ai\FakePendingDispatch;
 use Laravel\Ai\Jobs\GenerateEmbeddings;
+use Laravel\Ai\PendingResponses\Concerns\ResolvesProviderOptions;
 use Laravel\Ai\Prompts\QueuedEmbeddingsPrompt;
 use Laravel\Ai\Providers\Provider;
 use Laravel\Ai\Responses\Data\Meta;
 use Laravel\Ai\Responses\EmbeddingsResponse;
 use Laravel\Ai\Responses\QueuedEmbeddingsResponse;
-use Laravel\SerializableClosure\SerializableClosure;
 
 class PendingEmbeddingsGeneration
 {
-    use Conditionable;
+    use Conditionable, ResolvesProviderOptions;
 
     protected ?int $dimensions = null;
 
@@ -31,9 +30,6 @@ class PendingEmbeddingsGeneration
     protected ?bool $shouldCache = null;
 
     protected int $timeout = 30;
-
-    /** @var array<string, mixed>|SerializableClosure */
-    protected array|SerializableClosure $providerOptions = [];
 
     /**
      * Create a new pending embeddings generation instance.
@@ -95,38 +91,6 @@ class PendingEmbeddingsGeneration
         $this->timeout = $seconds;
 
         return $this;
-    }
-
-    /**
-     * Specify provider-specific options for embeddings generation.
-     *
-     * Pass a flat array to apply the same options to every selected provider,
-     * or a closure that receives the resolved Provider and returns the options
-     * for that provider. Queued closures may only capture serializable values.
-     *
-     * @param  array<string, mixed>|Closure(Provider): ?array<string, mixed>  $options
-     */
-    public function withProviderOptions(array|Closure $options): self
-    {
-        $this->providerOptions = $options instanceof Closure
-            ? new SerializableClosure($options)
-            : $options;
-
-        return $this;
-    }
-
-    /**
-     * Resolve provider options for the given provider.
-     *
-     * @return array<string, mixed>
-     */
-    protected function resolveProviderOptions(Provider $provider): array
-    {
-        if ($this->providerOptions instanceof SerializableClosure) {
-            return ($this->providerOptions)($provider) ?: [];
-        }
-
-        return $this->providerOptions;
     }
 
     /**
