+++ src/PendingResponses/PendingReranking.php
@@ -2,7 +2,6 @@
 
 namespace Laravel\Ai\PendingResponses;
 
-use Closure;
 use Illuminate\Support\Traits\Conditionable;
 use InvalidArgumentException;
 use Laravel\Ai\Ai;
@@ -11,16 +10,12 @@ use Laravel\Ai\Events\ProviderFailedOver;
 use Laravel\Ai\Exceptions\FailoverableException;
 use Laravel\Ai\Providers\Provider;
 use Laravel\Ai\Responses\RerankingResponse;
-use Laravel\SerializableClosure\SerializableClosure;
 
 class PendingReranking
 {
     use Conditionable;
 
     protected ?int $limit = null;
-
-    /** @var array<string, mixed>|SerializableClosure */
-    protected array|SerializableClosure $providerOptions = [];
 
     /**
      * Create a new pending reranking instance.
@@ -45,34 +40,6 @@ class PendingReranking
                 throw new InvalidArgumentException("Each document to rerank must be a non-blank string (index {$index}).");
             }
         }
-    }
-
-    /**
-     * Specify provider-specific options for reranking. Closures may only capture serializable values.
-     *
-     * @param  array<string, mixed>|Closure(Provider): ?array<string, mixed>  $options
-     */
-    public function withProviderOptions(array|Closure $options): self
-    {
-        $this->providerOptions = $options instanceof Closure
-            ? new SerializableClosure($options)
-            : $options;
-
-        return $this;
-    }
-
-    /**
-     * Resolve provider options for the given provider.
-     *
-     * @return array<string, mixed>
-     */
-    protected function resolveProviderOptions(Provider $provider): array
-    {
-        if ($this->providerOptions instanceof SerializableClosure) {
-            return ($this->providerOptions)($provider) ?: [];
-        }
-
-        return $this->providerOptions;
     }
 
     /**
